+++ fuocore/models.py
@@ -4,10 +4,32 @@
 fuocore.models
 ~~~~~~~~~~~~~~
 
-This modules defines several music models.
+这个模块定义了音乐资源的模型，如歌曲模型： ``SongModel`` , 歌手模型： ``ArtistModel`` 。
+它们都类似这样::
+
+    class XyzModel(BaseModel):
+        class Meta:
+            model_type = ModelType.xyz
+            fields = ['a', 'b', 'c']
+
+        @property
+        def ab(self):
+            return self.a + self.b
+
+同时，为了减少实现这些模型时带来的重复代码，这里还实现了：
+
+- ModelMeta: Model 元类，进行一些黑科技处理：比如解析 Model Meta 类
+- ModelMetadata: Model meta 属性对应的类
+- BaseModel: 基类
+
+ModelMetadata, ModelMeta, BaseModel 几个类是互相依赖的。
 """
 
 from enum import IntEnum
+import logging
+
+
+logger = logging.getLogger(__name__)
 
 
 class ModelType(IntEnum):
@@ -22,11 +44,20 @@ class ModelType(IntEnum):
     user = 17
 
 
+class ModelStage(IntEnum):
+    """Model 所处的阶段，有大小关系"""
+    display = 4
+    inited = 8
+    gotten = 16
+
+
 class ModelMetadata(object):
     def __init__(self,
                  model_type=ModelType.dummy.value,
                  provider=None,
                  fields=None,
+                 fields_display=None,
+                 fields_no_get=None,
                  allow_get=False,
                  allow_batch=False,
                  **kwargs):
@@ -37,34 +68,54 @@ class ModelMetadata(object):
         """
         self.model_type = model_type
         self.provider = provider
-        self.fields = fields
+        self.fields = fields or []
+        self.fields_display = fields_display or []
+        self.fields_no_get = fields_no_get or []
         self.allow_get = allow_get
         self.allow_batch = allow_batch
         for key, value in kwargs.items():
             setattr(self, key, value)
 
 
+class display_property:
+    """Model 的展示字段的描述器"""
+    def __init__(self, name):
+        #: display 属性对应的真正属性的名字
+        self.name_real = name
+        #: 用来存储值的属性名
+        self.store_pname = '_display_store_' + name
+
+    def __get__(self, instance, _=None):
+        if instance.stage >= ModelStage.inited:
+            return getattr(instance, self.name_real)
+        return getattr(instance, self.store_pname, '')
+
+    def __set__(self, instance, value):
+        setattr(instance, self.store_pname, value)
+
+
 class ModelMeta(type):
     def __new__(cls, name, bases, attrs):
-        # get all meta
+        # 获取 Model 当前以及父类中的 Meta 信息
+        # 如果 Meta 中相同字段的属性，子类的值可以覆盖父类的
         _metas = []
         for base in bases:
             base_meta = getattr(base, '_meta', None)
             if base_meta is not None:
                 _metas.append(base_meta)
-
-        # similar with djang/peewee model meta
         Meta = attrs.pop('Meta', None)
         if Meta:
             _metas.append(Meta)
 
-        fields = list()
-        meta_kv = {}
+        kind_fields_map = {'fields': [],
+                           'fields_display': [],
+                           'fields_no_get': []}
+        meta_kv = {}  # 实例化 ModelMetadata 的 kv 对
         for _meta in _metas:
-            inherited_fields = getattr(_meta, 'fields', [])
-            fields.extend(inherited_fields)
+            for kind, fields in kind_fields_map.items():
+                fields.extend(getattr(_meta, kind, []))
             for k, v in _meta.__dict__.items():
-                if k.startswith('_') or k in ('fields', ):
+                if k.startswith('_') or k in kind_fields_map:
                     continue
                 if k == 'model_type':
                     if ModelType(v) != ModelType.dummy:
@@ -79,12 +130,21 @@ class ModelMeta(type):
         model_type = meta_kv.pop('model_type', ModelType.dummy.value)
         if provider and ModelType(model_type) != ModelType.dummy:
             provider.set_model_cls(model_type, klass)
-        fields = list(set(fields))
+
+        fields_all = list(set(kind_fields_map['fields']))
+        fields_display = list(set(kind_fields_map['fields_display']))
+        fields_no_get = list(set(kind_fields_map['fields_no_get']))
+
+        for field in fields_display:
+            setattr(klass, field + '_display', display_property(field))
 
         # DEPRECATED attribute _meta
+        # TODO: remove this in verion 2.3
         klass._meta = ModelMetadata(model_type=model_type,
                                     provider=provider,
-                                    fields=fields,
+                                    fields=fields_all,
+                                    fields_display=fields_display,
+                                    fields_no_get=fields_no_get,
                                     **meta_kv)
         klass.source = provider.identifier if provider is not None else None
         # use meta attribute instead of _meta
@@ -92,17 +152,19 @@ class ModelMeta(type):
         return klass
 
 
-class Model(object, metaclass=ModelMeta):
+class Model(metaclass=ModelMeta):
     """base class for data models
+
     Usage::
 
         class User(Model):
             class Meta:
                 fields = ['name', 'title']
+
         user = UserModel(name='xxx')
         assert user.name == 'xxx'
         user2 = UserModel(user)
-        assert user2.name = 'xxx'
+        assert user2.name == 'xxx'
     """
 
     def __init__(self, obj=None, **kwargs):
@@ -118,7 +180,6 @@ class BaseModel(Model):
     """Base model for music resource.
 
     :param identifier: model object identifier, unique in each provider
-    :param source: model object provider identifier
 
     :cvar allow_get: meta var, whether model has a valid get method
     :cvar allow_list: meta var, whether model has a valid list method
@@ -128,23 +189,75 @@ class BaseModel(Model):
         allow_get = True
         allow_list = False
         model_type = ModelType.dummy.value
+
+        #: Model 所有字段，子类可以通过设置该字段以添加其它字段
         fields = ['identifier']
+
+        #: Model 用来展示的字段
+        fields_display = []
+
+        #: 不触发 get 的 Model 字段
+        fields_no_get = []
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
+
+        self.stage = kwargs.get('stage', ModelStage.inited)
 
     def __eq__(self, other):
         if not isinstance(other, BaseModel):
             return False
         return all([other.source == self.source,
                     other.identifier == self.identifier,
-                    other._meta.model_type == self._meta.model_type])
+                    other.meta.model_type == self.meta.model_type])
+
+    def __getattribute__(self, name):
+        """
+        获取 model 某一属性时，如果该属性值为 None 且该属性是 field
+        且该属性允许触发 get 方法，这时，我们尝试通过获取 model
+        详情来初始化这个字段，于此同时，还会重新给除 identifier
+        外的所 fields 重新赋值。
+        """
+        cls = type(self)
+        cls_name = cls.__name__
+        value = object.__getattribute__(self, name)
+        if name in cls.meta.fields \
+           and name not in cls.meta.fields_no_get \
+           and value is None \
+           and self.stage < ModelStage.gotten:
+            if cls.meta.allow_get:
+                logger.info("Model {} {}'s value is None, try to get detail."
+                            .format(repr(self), name))
+                obj = cls.get(self.identifier)
+                if obj is not None:
+                    for field in cls.meta.fields:
+                        if field in ('identifier', ):
+                            continue
+                        # 这里不能使用 getattr，否则有可能会无限 get
+                        fv = object.__getattribute__(obj, field)
+                        setattr(self, field, fv)
+                    self.stage = ModelStage.gotten
+                else:
+                    logger.warning('Model {} get return None'.format(cls_name))
+            else:
+                logger.warning("Model {} does't allow get".format(cls_name))
+            value = object.__getattribute__(self, name)
+        return value
+
+    @classmethod
+    def create_by_display(cls, identifier, **kwargs):
+        model = cls(identifier=identifier)
+        model.stage = ModelStage.display
+        for k, v in kwargs.items():
+            if k in cls.meta.fields_display:
+                setattr(model, k + '_display', v)
+        return model
 
     @classmethod
     def get(cls, identifier):
-        """Model get method
+        """获取 model 详情
 
-        Model should return a valid object if the identifier is available.
+        这个方法必须尽量初始化所有字段，确保它们的值不是 None。
         """
 
     @classmethod
@@ -215,7 +328,8 @@ class SongModel(BaseModel):
         model_type = ModelType.song.value
         # TODO: 支持低/中/高不同质量的音乐文件
         fields = ['album', 'artists', 'lyric', 'comments', 'title', 'url',
-                  'duration',]
+                  'duration']
+        fields_display = ['title', 'artists_name', 'album_name', 'duration_ms']
 
     @property
     def artists_name(self):
@@ -224,6 +338,13 @@ class SongModel(BaseModel):
     @property
     def album_name(self):
         return self.album.name if self.album is not None else ''
+
+    @property
+    def duration_ms(self):
+        if self.duration is not None:
+            seconds = self.duration / 1000
+            m, s = seconds / 60, seconds % 60
+        return '{:02}:{:02}'.format(int(m), int(s))
 
     @property
     def filename(self):