+++ fuocore/xiami/models.py
@@ -18,26 +18,11 @@ logger = logging.getLogger(__name__)
 
 
 class XBaseModel(BaseModel):
-    # FIXME: remove _detail_fields and _api to Meta
-    _detail_fields = ()
     _api = provider.api
 
     class Meta:
         allow_get = True
         provider = provider
-
-    def __getattribute__(self, name):
-        cls = type(self)
-        value = object.__getattribute__(self, name)
-        if name in cls._detail_fields and value is None:
-            logger.debug('Field %s value is None, get model detail first.' % name)
-            obj = cls.get(self.identifier)
-            for field in cls._detail_fields:
-                setattr(self, field, getattr(obj, field))
-            value = object.__getattribute__(self, name)
-        elif name in cls._detail_fields and not value:
-            logger.warning('Field %s value is not None, but is %s' % (name, value))
-        return value
 
 
 def _deserialize(data, schema_cls):
@@ -88,7 +73,6 @@ class XSongModel(SongModel, XBaseModel):
 
 
 class XAlbumModel(AlbumModel, XBaseModel):
-    _detail_fields = ('songs', 'artists', 'desc')
 
     @classmethod
     def get(cls, identifier):
@@ -99,7 +83,6 @@ class XAlbumModel(AlbumModel, XBaseModel):
 
 
 class XArtistModel(ArtistModel, XBaseModel):
-    _detail_fields = ('cover', 'desc')
 
     @classmethod
     def get(cls, identifier):
@@ -125,7 +108,6 @@ class XArtistModel(ArtistModel, XBaseModel):
 
 
 class XPlaylistModel(PlaylistModel, XBaseModel):
-    _detail_fields = ('songs', 'desc')
 
     class Meta:
         fields = ('uid', )