+++ tests/test_model.py
@@ -1,9 +1,10 @@
+from collections import namedtuple
 from unittest import TestCase
 
-from fuocore.models import Model
+from fuocore.models import Model, BaseModel, display_property
 
 
-class FakeProvider(object):
+class FakeProvider:
     identifier = 'fake'
     name = 'fake'
 
@@ -23,7 +24,7 @@ class TestModel(TestCase):
                 provider = provider
 
         song = SongModel()
-        self.assertEqual(song._meta.provider.name, 'fake')
+        self.assertEqual(song.meta.provider.name, 'fake')
 
     def test_meta_class_inherit(self):
         class SongModel(Model):
@@ -34,7 +35,7 @@ class TestModel(TestCase):
             pass
 
         song = LastSongModel()
-        self.assertEqual(song._meta.model_type, 1)
+        self.assertEqual(song.meta.model_type, 1)
 
     def test_meta_class_inherit_with_override(self):
         class SongModel(Model):
@@ -46,5 +47,43 @@ class TestModel(TestCase):
                 provider = provider
 
         song = LastSongModel()
-        self.assertEqual(song._meta.model_type, 1)
-        self.assertEqual(song._meta.provider.name, 'fake')
+        self.assertEqual(song.meta.model_type, 1)
+        self.assertEqual(song.meta.provider.name, 'fake')
+
+
+class TestBaseModel(TestCase):
+    def test_display_fields(self):
+        class SongModel(BaseModel):
+            class Meta:
+                fields = ['title', 'album']
+                fields_display = ['album_name']
+
+            @property
+            def album_name(self):
+                return self.album.name if self.album else ''
+
+        album_name = 'Minutes-to-Midnight'
+        song = SongModel.create_by_display(identifier=1, album_name=album_name)
+        self.assertEqual(song.album_name_display, album_name)
+        self.assertEqual(song.album_name, '')
+
+        real_album_name = 'Minutes to Midnight'
+        song.title = 'Leave out all the rest'
+        Album = namedtuple('Album', ('name', ))
+        song.album = Album(real_album_name)
+        song.use_display = False
+        self.assertEqual(song.album_name, real_album_name)
+
+
+class TestDisplayProperty(TestCase):
+    def test_display_basic_usage(self):
+        class A:
+            stage = 4
+            a_display = display_property('a')
+
+        a1 = A()
+        a2 = A()
+        self.assertEqual(a1.a_display, '')
+        a2.a_display = 'a2'
+        self.assertEqual(a1.a_display, '')
+        self.assertEqual(a2.a_display, 'a2')